klax_looper/requirements.txt
python-osc==1.8.0
python-rtmidi==1.5.8
